fix: ESLint 설정 수정 및 프로젝트 구조 재조정 (#24)
* MZ2-9 [fix]: eslint 관련 플러그인 및 설정 부분 수정

* MZ2-9 [fix]: commitlint 관련 라이브러리 제거

jira 와의 연동성을 위해 commitlint 를 제거하였습니다.

* MZ2-9 [fix]: eslint 관련 설정 일부 수정

불필요한 lint 설정을 대폭 삭제하고, 정확하게 lint 가 작동할 수 있도록 설정을 수정했습니다.

* MZ2-9 [feat]: git issue template 제거

JIRA를 사용하기로 결정하여 이슈 템플릿을 제거했습니다.

* MZ2-9 [fix]: 모든 파일에 대한 lint 수정 적용

모든 파일에 대해 새롭게 적용된 lint 룰을 적용하여 수정하였습니다.

* MZ2-9 [fix]: 컨벤션에 맞게 프로젝트 전면 재수정

사전에 협의한 컨벤션에 맞도록 프로젝트 컨벤션을 전면 재수정했습니다.
