+++ .eslintrc
@@ -1,9 +1,8 @@
 {
   "env": {
     "browser": true,
-    "node": true
+    "es6": true
   },
-  "parser": "@typescript-eslint/parser",
   "plugins": ["@typescript-eslint", "prettier"],
   "extends": [
     "airbnb",
@@ -11,6 +10,7 @@
     "plugin:@next/next/recommended",
     "prettier"
   ],
+  "parser": "@typescript-eslint/parser",
   "parserOptions": {
     "ecmaFeatures": {
       "jsx": true
@@ -21,38 +21,11 @@
   "settings": {
     "import/resolver": {
       "typescript": {
-        "directory": "./tsconfig.json"
+        "project": "./tsconfig.json"
       }
     }
   },
   "rules": {
-    "@typescript-eslint/no-empty-interface": "off",
-    "@typescript-eslint/no-unused-expressions": "off",
-    "@typescript-eslint/no-floating-promises": "off",
-    "@typescript-eslint/await-thenable": "off",
-    "@typescript-eslint/ban-ts-comment": "off",
-    "@typescript-eslint/no-unsafe-assignment": "off",
-    "@typescript-eslint/explicit-function-return-type": "off",
-    "@typescript-eslint/no-unsafe-return": "off",
-    "@typescript-eslint/no-unsafe-member-access": "off",
-    "@typescript-eslint/explicit-module-boundary-types": "off",
-    "@typescript-eslint/no-use-before-define": "off",
-    "@typescript-eslint/strict-boolean-expressions": "off",
-    "@typescript-eslint/unbound-method": "off",
-    "arrow-body-style": "off",
-    "consistent-return": "off",
-    "no-plusplus": "off",
-    "no-restricted-globals": "off",
-    "no-restricted-syntax": "off",
-    "no-nested-ternary": "off",
-    "no-unused-vars": "off",
-    "no-undef": "off",
-    "import/no-cycle": "off",
-    "import/extensions": ["off"],
-    "import/no-extraneous-dependencies": "off",
-    "import/no-default-export": "off",
-    "import/prefer-default-export": "off",
-    "react/button-has-type": "off",
     "react/jsx-filename-extension": [
       1,
       {
@@ -65,22 +38,15 @@
         "namedComponents": "arrow-function"
       }
     ],
-    "react/self-closing-comp": "error",
-    "react/no-unescaped-entities": "off",
-    "react/no-array-index-key": "off",
-    "react-hooks/exhaustive-deps": "warn",
-    "react-hooks/rules-of-hooks": "error",
-    "react/jsx-one-expression-per-line": "off",
-    "react/jsx-uses-react": "off",
-    "react/prop-types": "off",
     "react/react-in-jsx-scope": "off",
-    "react/require-default-props": [
-      "error",
-      {
-        "ignoreFunctionalComponents": true
-      }
-    ],
-    "react/jsx-wrap-multilines": "off",
-    "react/jsx-props-no-spreading": "off"
+    "react/button-has-type": "off",
+    "react/require-default-props": "off",
+    "react/jsx-props-no-spreading": "warn",
+    "react-hooks/exhaustive-deps": "warn",
+    "import/no-extraneous-dependencies": "off",
+    "import/extensions": "off",
+    "no-unused-vars": "warn",
+    "no-console": "off",
+    "import/prefer-default-export": "off"
   }
 }
