--- .github/ISSUE_TEMPLATE/epic.md
@@ -1,23 +1,0 @@
----
-name: Epic
-about: 새롭게 작업할 Epic을 작성합니다.
-title: ''
-labels: "\U0001F4D7 Epic"
-assignees: ''
-
----
-
-## 📚 Epic
-[작업할 Epic]
-
-## 🗂 Desc
-
-- 작업할 Epic 에 대한 설명 1
-- 작업할 Epic 에 대한 설명 2
-- 작업할 Epic 에 대한 설명 3
-
-## 🔍 Work
-
-- [x]  작업할 하위 Work 1
-- [x]  작업할 하위 Work 2
-- [x]  작업할 하위 Work 3
