+++ .lintstagedrc
@@ -1,3 +1,3 @@
 {
-  "src/**/*.[jt]s?(x)": "eslint --cache --fix"
+  "src/**/*.[jt]s?(x)": "eslint --fix"
 }
