--- commitlint.config.js
@@ -1,3 +1,0 @@
-const commitlintConfig = { extends: ['@commitlint/config-conventional'] };
-
-module.exports = commitlintConfig;
