+++ global.d.ts
@@ -1,4 +1,4 @@
-import React from 'react';
+import type React from 'react';
 
 declare module '*.mp3' {
   const src: string;
