+++ package.json
@@ -20,21 +20,17 @@
   },
   "devDependencies": {
     "@babel/core": "7.20.7",
-    "@commitlint/cli": "17.3.0",
-    "@commitlint/config-conventional": "17.3.0",
     "@next/eslint-plugin-next": "13.1.1",
     "@svgr/webpack": "6.5.1",
     "@tanstack/react-query-devtools": "4.20.4",
     "@types/node": "18.11.18",
     "@types/react": "18.0.26",
     "@types/react-dom": "18.0.10",
-    "@types/spotify-web-playback-sdk": "^0.1.15",
+    "@types/spotify-web-playback-sdk": "0.1.15",
     "@types/styled-components": "5.1.26",
     "@typescript-eslint/eslint-plugin": "5.47.1",
     "@typescript-eslint/parser": "5.47.1",
     "babel-loader": "9.1.0",
-    "babel-plugin-module-resolver": "4.1.0",
-    "commitlint": "17.3.0",
     "eslint": "8.31.0",
     "eslint-config-airbnb": "19.0.4",
     "eslint-config-next": "12.1.0",
