+++ src/apis/auth.ts
@@ -1,7 +1,7 @@
 import { postAsync } from '@/apis/API';
-import { AuthTokenType } from '@/stores/atoms';
-import { ApiResult } from '@/types/apiTypes';
-import { SocialType } from '@/types/authTypes';
+import type { AuthTokenType } from '@/stores/atoms';
+import type { ApiResult } from '@/types/apiTypes';
+import type { SocialType } from '@/types/authTypes';
 /**
  * 소셜 플랫폼으로부터 인계 받은 code를 넘겨 토큰을 받는 함수
  * @param social 인증을 진행한 소셜 플랫폼 타입
@@ -12,7 +12,7 @@ export const verifySocialAuthAsync = async (
   social: SocialType,
   code: string,
 ): ApiResult<AuthTokenType | null> => {
-  const response = await postAsync<AuthTokenType | null, any>(
+  const response = await postAsync<AuthTokenType | null, unknown>(
     `/users/login-tokens/oauth/${social}`,
     {
       code,
