--- src/components/Template/splash/disk/index.ts
@@ -1,3 +1,0 @@
-import DiskTemplate from './DiskTemplate';
-
-export default DiskTemplate;
