--- src/components/Template/splash/main/index.ts
@@ -1,3 +1,0 @@
-import MainTemplate from './MainTemplate';
-
-export default MainTemplate;
