--- src/components/Template/splash/matched/index.ts
@@ -1,3 +1,0 @@
-import MatchedTemplate from './MatchedTemplate';
-
-export default MatchedTemplate;
