+++ src/components/common/modal/modal-portal/ModalPortal.tsx
@@ -38,6 +38,6 @@ const ModalPortal = () => {
       : null;
   }
   return null;
-};
+}
 
 export default ModalPortal;
