+++ src/components/common/navigation/index.tsx
@@ -1,0 +1,3 @@
+import Navigation from './Navigation';
+
+export default Navigation;
