--- src/components/common/web-players/index.tsx
@@ -1,3 +1,0 @@
-import WebPlayers from './WebPlayers';
-
-export default WebPlayers;
