+++ src/components/layout/page-head/index.tsx
@@ -1,0 +1,3 @@
+import PageHead from './pageHead';
+
+export default PageHead;
