+++ src/components/layout/page-head/pageHead.tsx
@@ -1,0 +1,13 @@
+import Head from 'next/head';
+
+const PageHead = () => (
+  <Head>
+    <meta charSet="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <meta name="theme-color" content="#00917C" />
+    <link rel="icon" href="/favicon.ico" />
+    <title>MZ2MO : 음악의 세계를 열다</title>
+  </Head>
+);
+
+export default PageHead;
