+++ src/components/layout/page-layout/index.ts
@@ -1,0 +1,3 @@
+import PageLayout from './PageLayout';
+
+export default PageLayout;
