--- src/components/templates/BaseTemplates/index.ts
@@ -1,3 +1,0 @@
-import BaseTemplates from './BaseTemplates';
-
-export default BaseTemplates;
