+++ src/components/templates/splash/disk-template/SplashDiskTemplate.style.tsx
