+++ src/components/templates/splash/disk-template/index.ts
@@ -1,0 +1,3 @@
+import SplashDiskTemplate from './SplashDiskTemplate';
+
+export default SplashDiskTemplate;
