+++ src/components/templates/splash/main-template/SplashMainTemplate.style.tsx
