+++ src/components/templates/splash/main-template/index.ts
@@ -1,0 +1,3 @@
+import SplashMainTemplate from './SplashMainTemplate';
+
+export default SplashMainTemplate;
