+++ src/components/templates/splash/matched-template/index.ts
@@ -1,0 +1,3 @@
+import MatchedTemplate from './SplashMatchedTemplate';
+
+export default MatchedTemplate;
