+++ src/constants/apis/index.ts
@@ -1,1 +1,2 @@
 export { API_URL } from './server';
+export { SOCIAL_LOGIN_URL } from './auth';
