+++ src/constants/styles/theme.ts
@@ -1,4 +1,4 @@
-import { DefaultTheme } from 'styled-components';
+import type { DefaultTheme } from 'styled-components';
 
 const colors = {
   Text: {
