--- src/hooks/query/spotify/player/index.ts
@@ -1,1 +1,0 @@
-export { useEmojiPlaylist } from './useEmojiPlaylist';
