+++ src/pages/_app.tsx
@@ -9,6 +9,7 @@ import { GlobalStyle } from '@/constants/styles/globalStyle';
 import { theme } from '@/constants/styles/theme';
 import '@/assets/fonts/font.css';
 
+import PageHead from '@/components/layout/page-head';
 import ModalPortal from '@/components/common/modal/modal-portal';
 
 const MyApp = ({ Component, pageProps }: AppProps) => {
@@ -29,7 +30,9 @@ const MyApp = ({ Component, pageProps }: AppProps) => {
       <Provider>
         <GlobalStyle />
         <ThemeProvider theme={theme}>
+          <PageHead />
           <ModalPortal />
+          {/* eslint-disable react/jsx-props-no-spreading */}
           <Component {...pageProps} />
         </ThemeProvider>
       </Provider>
