+++ src/pages/splash/match.tsx
@@ -1,12 +1,10 @@
-import BaseTemplates from '@/components/templates/BaseTemplates';
-import MatchedTemplate from '@/components/Template/splash/matched';
+import PageLayout from '@/components/layout/page-layout';
+import SplashMatchedTemplate from '@/components/templates/splash/matched-template';
 
-const Match = () => {
-  return (
-    <BaseTemplates>
-      <MatchedTemplate />
-    </BaseTemplates>
-  );
-};
+const Match = () => (
+  <PageLayout>
+    <SplashMatchedTemplate />
+  </PageLayout>
+);
 
 export default Match;
