+++ src/stores/atoms.ts
@@ -1,4 +1,4 @@
-import { ReactNode } from 'react';
+import type { ReactNode } from 'react';
 import { atom } from 'jotai';
 import { atomWithStorage } from 'jotai/utils';
